pyproject.toml
[project]
name = "pool-backend"
version = "0.1.0"
description = "Pool backend API with FastAPI"
authors = [
    {name = "Romain Dardour"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "email-validator (>=2.1.0,<3.0.0)",
    "letta-client (>=0.1.0,<0.2.0)"
]

[tool.poetry]
name = "pool-backend"
version = "0.1.0"
description = "Pool backend API with FastAPI"
authors = ["Romain Dardour"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.13"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
python-multipart = "^0.0.20"
python-jose = "^3.5.0"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
sqlalchemy = "^2.0.41"
alembic = "^1.16.2"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"
email-validator = "^2.1.0"
letta-client = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "sqlalchemy.*",
]
ignore_missing_imports = true
